backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv

sqlalchemy
psycopg2-binary
alembic

passlib[argon2]
python-jose[cryptography]
email-validator

chromadb
openai
tiktoken

langchain
langchain-community
langchain-openai
langchain-text-splitters
requests

pymupdf
transformers
torch
Pillow
pytesseract
scikit-learn
faiss-cpu  